Extract longitude and latitude values

Extracts longitude and latitude values for a particular region from the world map supplied by the maps package.

Usage

mapdata(region = ".", map = "world", exact = FALSE, ...)

Arguments

region

Passed to map as the argument regions.

map

Passed to map as the argument database

exact

The argument exact passed to the map function.

...

Additional arguments to be passed to map.

Value

A dataframe with two columns: long and lat for longitude and latitude.

Examples

See the examples in mimic.