Print method for objects of class "anscombe"

print method for class "anscombe".

Usage

# S3 method for class 'anscombe'
print(x, ...)

Arguments

x

an object of class "anscombe", a result of a call to anscombise or mimic.

...

Additional optional arguments to be passed to print.

Value

The argument x, invisibly.

Details

Just extracts the new dataset from x and prints it using print.

See also